+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import MainImage from "../../public/main-img.svg";
 import Image from 'next/image';
-import Header from "@/components/molecules/header";
-import Footer from "@/components/molecules/footer";
+import Header from "@/components/molecules/Header/header";
+import Footer from "@/components/molecules/Footer/footer";
 
 import PublicOnlyFeature from "@/components/templates/Public/public";
 
@@ -11,7 +11,7 @@ export default function Home() {
       <main className="bg-[#FFEAD4] overflow-hidden h-screen">
         <div className="flex flex-col items-center justify-center gap-3 h-full">
           <Header />
-          <Image src={MainImage} alt="MainImage" className="mb-56"/>
+          <Image src={MainImage} alt="MainImage" className="mb-56" />
           <Footer />
         </div>
       </main>
